extract common TLS test setup
