+++ rust-runtime/aws-smithy-http-client/tests/common/mod.rs
@@ -6,3 +6,6 @@
 //! Shared support for `aws-smithy-http-client` integration tests.
 
 pub(crate) mod client;
+
+#[cfg(any(feature = "__rustls", feature = "s2n-tls"))]
+pub(crate) mod tls;
